Cleanup of script names
