--- Game/Assets/Prefabs/NPCWAVES/New Sprite.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1daa24285a1bc41609c257da221937c1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
