+++ Game/Assets/Prefabs/UI/Dialogue Box.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a8bb7b462f0101846a8f3fe081bbe0ac
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
