+++ Game/Assets/Scripts/Interactive Objects/ClockoutMachineInteractManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3278b39d68627434bafee9b34548b1b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
