+++ Game/Assets/Scripts/Interactive Objects/CoffeeMachineInteractManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e512af8cd352f14d8542ae633c0b0de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
