+++ Game/Assets/Scripts/Interactive Objects/DoorInteractManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a20a9effecdbf542b517d16638dab9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
