+++ Game/Assets/Scripts/Interactive Objects/NPCInteractManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 95630db81685ee34a82aae7d14faec82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
