--- Game/Assets/Scripts/Interactive Objects/Patient.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 95630db81685ee34a82aae7d14faec82
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
