--- Game/Assets/Scripts/Interactive Objects/TV.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4244e539b1f8b7c4a993d184972b3d95
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
