+++ Game/Assets/Scripts/Interfaces/IInteractable.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IInteractable
+{
+    void Interact();
+    void ManualHighlight();
+}
