--- Game/Assets/Scripts/Interfaces/IInteractive.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IInteractive
-{
-    void Interact();
-    void ManualHighlight();
-}
