--- Game/Assets/Scripts/Interfaces/IInteractive.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 640c37741efe4444fbb62125cc0ef4fd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
