+++ Game/Assets/Scripts/Managers/DialogueManager.cs
@@ -86,7 +86,7 @@ public class DialogueManager : MonoBehaviour
     public void EndDialogue()
     {
         animator.SetBool("IsOpen", false);
-        GetComponent<Patient>().isTalking = false;
+        GetComponent<NPCInteractManager>().isTalking = false;
         if (dialogue.GetNextDialogue().Length == 0 && diagnosisDone)
         {
             dialogue = diagnosisDoneDialogue;
