--- Game/Assets/Scripts/UI/New Dialogue.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: feff62ab79152e94a9a0cadff3e1b955
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
